chore: move HomePage to (home) folder
