+++ apps/scrum/src/app/(home)/BubbleBackground.tsx
