+++ apps/scrum/src/app/(home)/page.tsx
@@ -1,10 +1,10 @@
 import Image from 'next/image'
 
 import Link from 'next/link'
-import fore from '../public/backgrounds/fore.png'
-import upper from '../public/backgrounds/upper.png'
-import back from '../public/backgrounds/back.png'
-import tica from '../public/icons/tica.png'
+import fore from '../../public/backgrounds/fore.png'
+import upper from '../../public/backgrounds/upper.png'
+import back from '../../public/backgrounds/back.png'
+import tica from '../../public/icons/tica.png'
 import ConversationArticle from './ConversationArticle'
 import ScrollButton from './ScrollButton'
 import BubbleBackground from './BubbleBackground'
