+++ apps/scrum/src/app/(scrum)/flows/page.tsx
@@ -3,7 +3,7 @@ import { ScrumRoute } from '../constants'
 import rock from '@/public/backgrounds/rock.png'
 import ScrumHeader from '../ScrumHeader'
 import ScrumNav from '../ScrumNav'
-import ConversationArticle from '@/app/ConversationArticle'
+import ConversationArticle from '@/app/(home)/ConversationArticle'
 import DevelopmentTeamIcon from '../characters/DevelopmentTeamIcon'
 import DiagramBackground from './DiagramBackground'
 import LongArrowIcon from './LongArrowIcon'
