+++ apps/scrum/src/app/(scrum)/sprints/page.tsx
@@ -3,7 +3,7 @@ import { NORMALISED_TABS, ScrumRoute } from '../constants'
 import camp from '@/public/backgrounds/camp.png'
 import ScrumHeader from '../ScrumHeader'
 import ScrumNav from '../ScrumNav'
-import ConversationArticle from '@/app/ConversationArticle'
+import ConversationArticle from '@/app/(home)/ConversationArticle'
 import DevelopmentTeamIcon from '../characters/DevelopmentTeamIcon'
 import SprintTab from './SprintTab'
